chore: update color, grid and space scheme 1.1.0

+++ ouds_theme_contract/lib/theme/scheme/responsive/ouds_grid_scheme.dart
@@ -59,7 +59,7 @@ class OudsGridScheme {
         medium: gridTokens.mediumMargin,
       );
 
-  double get columnGap => OudsWindowSizeClassUtil.select(
+  int get columnGap => OudsWindowSizeClassUtil.select(
         sizeClass: sizeClass,
         extraCompact: gridTokens.extraCompactColumnGap,
         compact: gridTokens.compactColumnGap,
